+++ module/workflow/src/Infrastructure/Handler/Status/DeleteStatusCommandHandler.php
@@ -9,7 +9,10 @@ declare(strict_types = 1);
 
 namespace Ergonode\Workflow\Infrastructure\Handler\Status;
 
+use Ergonode\Core\Infrastructure\Exception\ExistingRelationshipsException;
+use Ergonode\Core\Infrastructure\Resolver\RelationshipsResolverInterface;
 use Ergonode\Workflow\Domain\Command\Status\DeleteStatusCommand;
+use Ergonode\Workflow\Domain\Entity\Status;
 use Ergonode\Workflow\Domain\Repository\StatusRepositoryInterface;
 use Webmozart\Assert\Assert;
 
@@ -23,22 +26,37 @@ class DeleteStatusCommandHandler
     private $repository;
 
     /**
-     * @param StatusRepositoryInterface $repository
+     * @var RelationshipsResolverInterface
      */
-    public function __construct(StatusRepositoryInterface $repository)
-    {
+    private $relationshipsResolver;
+
+    /**
+     * @param StatusRepositoryInterface      $repository
+     * @param RelationshipsResolverInterface $relationshipsResolver
+     */
+    public function __construct(
+        StatusRepositoryInterface $repository,
+        RelationshipsResolverInterface $relationshipsResolver
+    ) {
         $this->repository = $repository;
+        $this->relationshipsResolver = $relationshipsResolver;
     }
 
     /**
      * @param DeleteStatusCommand $command
      *
+     * @throws ExistingRelationshipsException
      * @throws \ReflectionException
      */
     public function __invoke(DeleteStatusCommand $command)
     {
         $status = $this->repository->load($command->getId());
-        Assert::notNull($status);
+        Assert::isInstanceOf($status, Status::class, sprintf('Can\'t find status with ID "%s"', $command->getId()));
+
+        $relationships = $this->relationshipsResolver->resolve($command->getId());
+        if (!$relationships->isEmpty()) {
+            throw new ExistingRelationshipsException($command->getId());
+        }
 
         $this->repository->delete($status);
     }
